+++ usb2canfdv1-fw/Slcan/Src/generator.c
@@ -32,16 +32,16 @@ enum SlcanTimestampMode slcan_timestamp_mode = 0;
 uint16_t slcan_report_reg = 1;   // Default: no timestamp, no ESI, no Tx, but with Rx
 
 // Private methods
-static int32_t slcan_generate_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_header, uint8_t *frame_data);
+static uint16_t slcan_generate_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_header, uint8_t *frame_data);
 
 // Generate a slcan message from a CAN frame
 // Returns number of bytes written into buf
 //  MIN: 1 (r) + SLCAN_STD_ID_LEN + 2 (DLC & [CR])
 //  MAX: SLCAN_MTU - 1 (z/Z) - 16 (padding)
-int32_t slcan_generate_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_header, uint8_t *frame_data)
+uint16_t slcan_generate_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_header, uint8_t *frame_data)
 {
     // Start building the slcan message string at idx 0 in buf
-    uint8_t msg_idx = 0;
+    uint16_t msg_idx = 0;
 
     // Handle remote frames
     if (frame_header->RxFrameType == FDCAN_REMOTE_FRAME)
@@ -153,7 +153,7 @@ int32_t slcan_generate_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_header, 
 // Returns number of bytes written into buf
 //  MIN: 1 (r) + SLCAN_STD_ID_LEN + 2 (DLC & [CR])
 //  MAX: SLCAN_MTU - 1 (z/Z) - 16 (padding)
-int32_t slcan_generate_rx_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_header, uint8_t *frame_data)
+uint16_t slcan_generate_rx_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_header, uint8_t *frame_data)
 {
     // Check if Rx reporting is required
     if (((slcan_report_reg >> SLCAN_REPORT_RX) & 1) == 0)
@@ -162,7 +162,7 @@ int32_t slcan_generate_rx_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_heade
     if (buf == NULL)
         return 0;
 
-    int32_t len = slcan_generate_frame(buf, frame_header, frame_data);
+    uint16_t len = slcan_generate_frame(buf, frame_header, frame_data);
 
     // Return string length
     return len;
@@ -172,7 +172,7 @@ int32_t slcan_generate_rx_frame(uint8_t *buf, FDCAN_RxHeaderTypeDef *frame_heade
 // Returns number of bytes written into buf
 //  MIN: 1 (r) + SLCAN_STD_ID_LEN + 2 (DLC & [CR])
 //  MAX: SLCAN_MTU - 16 (padding)
-int32_t slcan_generate_tx_event(uint8_t *buf, FDCAN_TxEventFifoTypeDef *tx_event, uint8_t *frame_data)
+uint16_t slcan_generate_tx_event(uint8_t *buf, FDCAN_TxEventFifoTypeDef *tx_event, uint8_t *frame_data)
 {
     // Check if Tx reporting is required
     if (((slcan_report_reg >> SLCAN_REPORT_TX) & 1) == 0)
@@ -195,7 +195,7 @@ int32_t slcan_generate_tx_event(uint8_t *buf, FDCAN_TxEventFifoTypeDef *tx_event
     frame_header.BitRateSwitch = tx_event->BitRateSwitch;
     frame_header.FDFormat = tx_event->FDFormat;
     frame_header.RxTimestamp = tx_event->TxTimestamp;
-    int32_t len = slcan_generate_frame(&buf[1], &frame_header, frame_data);
+    uint16_t len = slcan_generate_frame(&buf[1], &frame_header, frame_data);
 
     // Return string length
     return len + 1;
